requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
altair==4.2.2
annotated-types==0.5.0
ansi2html==1.8.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autopep8==1.7.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
build==1.0.3
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
comtypes==1.1.14
contourpy==1.0.7
cycler==0.11.0
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.7
dataclasses-json==0.5.14
datasets==2.14.4
DateTime==4.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.103.0
fastjsonschema==2.17.1
ffmpy==0.3.1
filelock==3.12.2
FlashRank==0.1.66
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
gradio==3.23.0
gradio_client==0.5.0
greenlet==2.0.2
grpcio==1.60.0
grpcio-status==1.48.2
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.0.1
inflate64==0.3.1
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.17
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.4.4
kubernetes==29.0.0
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langchain-experimental==0.0.25
langsmith==0.0.83
libclang==16.0.0
linkify-it-py==2.0.3
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
mlflow==2.6.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
mtcnn==0.1.0
multidict==6.0.4
multiprocess==0.70.15
multivolumefile==0.2.3
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.8
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.16.3
openai==0.28.0
opencv-python==4.8.0.76
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
orjson==3.9.5
overrides==7.3.1
packaging==23.2
panda==0.3.1
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pdf2image==1.17.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.8.1
playsound==1.3.0
plotly==5.18.0
portalocker==2.7.0
posthog==3.3.2
prometheus-client==0.17.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==3.19.6
psutil==5.9.5
pulsar-client==3.4.0
pure-eval==0.2.2
py7zr==0.20.6
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
PyMuPDF==1.23.23
PyMuPDFb==1.23.22
PyMySQL==1.1.0
pyparsing==3.0.9
pypdf==4.0.2
PyPDF2==3.0.1
PyPika==0.48.9
pypiwin32==223
pyppmd==1.0.0
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pyttsx3==2.90
pytz==2022.4
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
pyzstd==0.15.9
querystring-parser==1.2.4
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rouge-score==0.1.2
rsa==4.9
sacrebleu==2.3.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.0
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
streamlit==1.22.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-addons==0.21.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typeguard==2.13.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==1.26.14
uvicorn==0.23.2
validators==0.20.0
waitress==2.1.2
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.14.1
xgboost==1.7.6
xxhash==3.3.0
yarl==1.9.2
youtube-transcript-api==0.6.2
zipp==3.15.0
zope.interface==5.5.0
